# server/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MediaTracker",
    "version": "0.2.11",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "security": [
    {
      "ApiKey": []
    },
    {
      "ApiHeader": []
    }
  ],
  "paths": {
...
    "/api/discover/trending": {
      "get": {
        "tags": [
          "Discover"
        ],
        "operationId": "trending",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrendingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/discover/add-to-watchlist": {
      "post": {
        "tags": [
          "Discover"
        ],
        "operationId": "addToWatchlist",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tmdbId": {
                    "type": "number"
                  }
                },
                "required": [
                  "tmdbId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mediaItemId": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "mediaItemId"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/details/update-metadata/{mediaItemId}": {
      "get": {
        "tags": [
          "MediaItem"
        ],
        "operationId": "updateMetadata",
        "parameters": [
          {
            "name": "mediaItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/text": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "query",
        "name": "token"
      },
      "ApiHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    },
    "schemas": {
...
      "TrendingResponse": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrendingTvShow"
            }
          },
          "page": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "results",
          "totalPages"
        ]
      },
      "TrendingTvShow": {
        "type": "object",
        "properties": {
          "tmdbId": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "posterPath": {
            "type": "string",
            "nullable": true
          },
          "backdropPath": {
            "type": "string",
            "nullable": true
          },
          "firstAirDate": {
            "type": "string",
            "nullable": true
          },
          "tmdbRating": {
            "type": "number"
          },
          "isOnWatchlist": {
            "type": "boolean"
          },
          "mediaItemId": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "isOnWatchlist",
          "overview",
          "title",
          "tmdbId",
          "tmdbRating"
        ]
      },
...
    }
  }
}